Dispatch addSeenGuidedTour from quitGuidedTour, not from component
The component shouldn't have to care about `addSeenGuidedTour`, which is
IMO an implementation detail; from its name, dispatching
`quitGuidedTour` should be more than enough, so let that action actually
dispatch two.

The downside is that `quitGuidedTour` is no longer defined as a pure
action creator, but as a thunk creator.